fix: extension redirect
