+++ packages/extension/src/newtab/MainFeedPage.tsx
@@ -1,4 +1,10 @@
-import React, { ReactElement, useContext, useMemo, useState } from 'react';
+import React, {
+  ReactElement,
+  useContext,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react';
 import usePersistentContext from '@dailydotdev/shared/src/hooks/usePersistentContext';
 import MainLayout from '@dailydotdev/shared/src/components/MainLayout';
 import MainFeedLayout, {
@@ -10,6 +16,7 @@ import TimerIcon from '@dailydotdev/shared/icons/timer.svg';
 import AuthContext from '@dailydotdev/shared/src/contexts/AuthContext';
 import SimpleTooltip from '@dailydotdev/shared/src/components/tooltips/SimpleTooltip';
 import { HeaderButton } from '@dailydotdev/shared/src/components/buttons/common';
+import { useMyFeed } from '@dailydotdev/shared/src/hooks/useMyFeed';
 import MostVisitedSites from './MostVisitedSites';
 
 const PostsSearch = dynamic(
@@ -35,6 +42,7 @@ export default function MainFeedPage({
   const [isSearchOn, setIsSearchOn] = useState(false);
   const [searchQuery, setSearchQuery] = useState<string>();
   const [showDnd, setShowDnd] = useState(false);
+  const { registerLocalFilters } = useMyFeed();
   const [defaultFeed] = usePersistentContext('defaultFeed', 'popular');
   const enableSearch = () => {
     setIsSearchOn(true);
@@ -69,6 +77,19 @@ export default function MainFeedPage({
     setIsSearchOn(false);
     setSearchQuery(undefined);
   };
+
+  useEffect(() => {
+    const urlParams = new URLSearchParams(window.location.search);
+    const createFilter = urlParams.get('create_filter');
+
+    if (createFilter) {
+      registerLocalFilters().then(({ hasFilters }) => {
+        if (hasFilters) {
+          setFeedName('my-feed');
+        }
+      });
+    }
+  }, []);
 
   return (
     <MainLayout
